requirements.txt
apiclient==1.0.4
aspy.yaml==1.3.0
atomicwrites==1.3.0
attrs==19.3.0
bleach==3.1.0
build-utils==0.3.2
cachetools==3.1.1
certifi==2019.9.11
cfgv==2.0.1
chardet==3.0.4
colorama==0.4.1
docutils==0.15.2
fake-rpi==0.6.2
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
httplib2==0.14.0
identify==1.4.7
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
nodeenv==1.3.3
nose==1.3.7
numpy==1.17.3
oauth2client==4.1.3
packaging==19.2
paho-mqtt==1.4.0
pkginfo==1.5.0.1
pluggy==0.13.0
pre-commit==1.18.3
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.4.2
PyGObject==3.34.0
pyparsing==2.4.2
pytest==5.2.1
PyYAML==5.1.2
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.12.0
toml==0.10.0
tqdm==4.36.1
twine==2.0.0
uritemplate==3.0.0
urllib3==1.25.6
virtualenv==16.7.5
wcwidth==0.1.7
webencodings==0.5.1
zipp==0.6.0
